stupid casting stuff

+++ Project1/Search.h
@@ -73,7 +73,7 @@ ResizableArray<long long>* search(const T& item, const ResizableArray<T>& array,
 		idxArray->add(idx);
 
 		// Searches indexes after the returned index to see if multiple items exist
-		while ((idx >= 0) && (idx < array.getSize()) && ((comparator.compare(item, array.get((unsigned long)(idx + 1)))) == 0)) {
+		while ((idx >= 0) && (idx < (long long)(array.getSize())) && ((comparator.compare(item, array.get((unsigned long)(idx + 1)))) == 0)) {
 			idxArray->add(idx + 1);
 			idx++;
 		}
